cdk/requirements-cdk.txt
aws-cdk-lib==2.118.0
constructs>=10.0.0